check review adiciona config individual + application.proprieties
