+++ src/main/resources/application.properties
@@ -26,5 +26,14 @@ spring.mail.properties.mail.smtp.starttls.enable=true
 spring.rabbitmq.addresses= amqps://wggzhuka:X0WbwZffhRrQ5HztDS3SJmA1uIYzCjHw@beaver.rmq.cloudamqp.com/wggzhuka
 spring.rabbitmq.queue=ms.email
 
+# Swagger config
 refpath.swagger.path=localhost:8080
 spring.mvc.pathmatch.matching-strategy = ANT_PATH_MATCHER
+
+refpath.swagger.apiinfo.title: "Refinaria API"
+refpath.swagger.apiinfo.description: "Sistema de refinaria para nivelamento de candidatos"
+refpath.swagger.apiinfo.termsOfServiceUrl: "http://localhost:8080"
+refpath.swagger.apiinfo.license: ""
+refpath.swagger.apiinfo.licenseUrl: "http://localhost:8080"
+refpath.swagger.apiinfo.version: "2.0"
+
